added -Wall flag
added the -Wall flag and corrected everything it complained about.

the strict aliasing bug still persists in Testing/run.asm.
i'd hoped this would fix it (or show me where it was), but apparently compilers aren't very good at detecting aliasing issues.
however, i have a hunch where the issue is and will be applying a potential patch in the next commit.

+++ Interface.cpp
@@ -85,7 +85,7 @@ namespace CSX64
 			// make room for the pointer array
 			stack -= 8 * pointers.size();
 			// write pointer array to memory
-			for (int i = 0; i < pointers.size(); ++i) SetMem(stack + i * 8, pointers[i]);
+			for (int i = 0; i < (int)pointers.size(); ++i) SetMem(stack + i * 8, pointers[i]);
 		}
 
 		// load arg count and arg array pointer to RDI, RSI
